feat(modularization): Phase 4 in-process GraphQL execute in clawql-api
Move graphql schema builder, execute helpers, and in-process execute into
clawql-api; execute-core calls them directly. ExecuteEnvironment now only
injects native GraphQL/gRPC IO from transport. Add @omnigraph/openapi deps.
