+++ packages/clawql-api/package.json
@@ -23,8 +23,10 @@
     "test": "vitest run"
   },
   "dependencies": {
+    "@graphql-mesh/utils": "^0.104.36",
     "@grpc/grpc-js": "^1.14.4",
     "@grpc/proto-loader": "^0.8.1",
+    "@omnigraph/openapi": "^0.109.51",
     "clawql-core": "file:../clawql-core",
     "effect": "3.21.2",
     "graphql": "^16.14.0",
